+++ packages/@tinacms/toolkit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.7.8
+
+### Patch Changes
+
+- f6efd498e: Added branch creation model
+
 ## 1.7.7
 
 ### Patch Changes